modern opengl DONE
- fix samples 005-010
- fix KgtAssetManager not exposing the correct getImageSize API
- fix a bunch of KGT code-utilities
- fix some minor stuff in kmath
- re-implement krbSetWireframe

+++ code/kmath.cpp
@@ -1159,7 +1159,7 @@ internal inline void kmath::generateMeshBox(
 		*V2F32_STRIDE(o_texNormals,vertexByteStride,v) = TRI_TEX_NORMS[v];
 	}
 }
-internal inline size_t kmath::generateMeshCircleSphereVertexCount(
+internal inline u32 kmath::generateMeshCircleSphereVertexCount(
 		u32 latitudeSegments, u32 longitudeSegments)
 {
 	return ((latitudeSegments - 2)*6 + 6) * longitudeSegments;
